Extract TokenManager class to its own module
